cleanup

+++ Rakefile
@@ -5,12 +5,13 @@ require 'jekyll'
 REPONAME = "ranmocy/ranmocy"
 ROOT_DIR  = Pathname.new('.').expand_path
 POSTS_DIR = Pathname.new("_posts").expand_path
+SITE_DIR = Pathname.new("_site").expand_path
 
 YAMLFrontMatter =<<FRONT
 ---
 layout: default
 published: true
-date: 22 January 2013
+date: 11 May 1991
 category: %s
 tags: [%s]
 ---
@@ -21,7 +22,8 @@ FRONT
 namespace :site do
   desc "Cleanup"
   task :cleanup do
-    FileUtils.rm_r(POSTS_DIR)
+    FileUtils.rm_r(SITE_DIR) if SITE_DIR.exist?
+    FileUtils.rm_r(POSTS_DIR) if POSTS_DIR.exist?
     FileUtils.mkdir(POSTS_DIR)
   end
 
@@ -29,7 +31,7 @@ namespace :site do
   task :prepare => [:cleanup] do
     Dir["[^_]*/*.*"].each do |file|
       ord_file = ROOT_DIR.join(file)
-      filename = ord_file.basename
+      filename = ord_file.basename.to_s.gsub(/\s+/, '-')
       category = ord_file.dirname.basename
       new_file = POSTS_DIR.join("2013-04-13-#{filename}")
